+++ paper_uploads/forms/dialogs/base.py
@@ -1,6 +1,6 @@
 from django import forms
 from django.utils.translation import gettext_lazy as _
-from paper_admin.renderer import PaperFormRenderer
+from paper_admin.admin.renderers import PaperFormRenderer
 
 
 class UploadedFileBaseForm(forms.ModelForm):